more changes done 6 Aug
