+++ src/app/auth/auth.interceptor.ts
@@ -11,7 +11,7 @@ export class AuthInterceptor implements HttpInterceptor {
         
     }
     intercept(req: HttpRequest<any>, next: HttpHandler): Observable<HttpEvent<any>> {
-        if(localStorage.getItem('accessToken')!=null && localStorage.getItem("refreshToken")) {
+        if(localStorage.getItem('accessToken')!=null && localStorage.getItem("refreshToken")!=null) {
             const clonedReq = req.clone({
                 headers : req.headers.set('Authorization','Bearer '+localStorage.getItem('accessToken'))
             });
