+++ src/app/header/header.component.html
@@ -21,7 +21,7 @@
       </li>
       <li class="nav-item ">
         <a class="nav-link" routerLink="/cart">
-          <span #cartItemsCount class="mat-card mat-elevation-z4 mat-badge-hidden" matBadge="{{ this.count }}"
+          <span #cartItemsCount class="mat-card mat-elevation-z4 mat-badge-hidden" matBadge="{{ this.bookCount }}"
             matBadgeOverlap="false" matBadgeColor="accent">Cart</span>
         </a>
       </li>
